feat: Support JSON Schema drafts 2019-09 and 2020-12 and JSON Type Definition
Upgrade AJV to the latest version and retain the previous AJV@6
to be able to suport JSON Schema draft 04.

BREAKING CHANGE: The default environment recognises only JSON Schema drafts 06 and 07 automatically. Not 04 any more. The environment for JSON Schema drafts 04 has to be selected explicitly. Also, JSON Schema drafts 06 and 07 are handled by AJV@8 instead of AJV@6. It shouldn't make any difference, but the implementation is new and could perform a stricter validation.